Move HFunctor & Effect into their own module.
