+++ src/Control/Effect/Pure.hs
@@ -4,7 +4,11 @@ module Control.Effect.Pure
   Pure
 ) where
 
+import Control.Effect.Class
 import GHC.Generics (Generic1)
 
 data Pure (m :: * -> *) k
   deriving (Functor, Generic1)
+
+instance HFunctor Pure
+instance Effect   Pure
